third_party/blink/web_tests/fast/performance/performance-mark-exceptions-expected.txt

This tests that 'performance.mark' throws exceptions with reasonable messages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.performance.mark('navigationStart') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'navigationStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('unloadEventStart') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'unloadEventStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('unloadEventEnd') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'unloadEventEnd' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('redirectStart') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'redirectStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('redirectEnd') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'redirectEnd' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('fetchStart') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'fetchStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('domainLookupStart') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'domainLookupStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('domainLookupEnd') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'domainLookupEnd' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('connectStart') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'connectStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('connectEnd') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'connectEnd' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('secureConnectionStart') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'secureConnectionStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('requestStart') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'requestStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('responseStart') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'responseStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('responseEnd') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'responseEnd' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('domLoading') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'domLoading' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('domInteractive') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'domInteractive' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('domContentLoadedEventStart') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'domContentLoadedEventStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('domContentLoadedEventEnd') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'domContentLoadedEventEnd' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('domComplete') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'domComplete' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('loadEventStart') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'loadEventStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('loadEventEnd') threw exception SyntaxError: Failed to execute 'mark' on 'Performance': 'loadEventEnd' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS successfullyParsed is true

TEST COMPLETE
